hwloc: use int in hwloc_bitmap_foreach_begin loop
Affinity.c:67:10: runtime error: implicit conversion from type 'int' of value -1 (32-bit, signed) to type 'unsigned int' changed the value to 4294967295 (32-bit, unsigned)

+++ AffinityPanel.c
@@ -425,9 +425,9 @@ Affinity* AffinityPanel_getAffinity(Panel* super, ProcessList* pl) {
    Affinity* affinity = Affinity_new(pl);
 
    #ifdef HAVE_LIBHWLOC
-   unsigned int i;
+   int i;
    hwloc_bitmap_foreach_begin(i, this->workCpuset)
-   Affinity_add(affinity, i);
+      Affinity_add(affinity, (unsigned)i);
    hwloc_bitmap_foreach_end();
    #else
    for (int i = 0; i < Vector_size(this->cpuids); i++) {
